use kea for product deets since it needs to be used elsewhere
