+++ .gitignore
@@ -31,3 +31,4 @@ fonts
 src/components/Layout/Fonts.scss
 .vercel
 .env
+*Type.ts
